requirements.txt
openai>=1.0
anthropic>=0.30
langchain>=0.2
langchain-openai>=0.1
langchain-anthropic>=0.1
crewai>=0.30
python-dotenv>=1.0
rich>=13.0
pydantic>=2.0
httpx>=0.25
